chore: set up submodule splitting
