+++ .config/rollup.config.js
@@ -8,30 +8,55 @@ import resolve from '@rollup/plugin-node-resolve';
 import pkg from '../package.json' assert { type: 'json' };
 
 const bundle = (config) => ({
-    input: './src/index.ts',
     external: (id) => !/^[./]/.test(id),
     ...config,
 });
 
 export default [
     bundle({
+        input: './src/index.ts',
         plugins: [commonjs(), resolve(), esbuild(), cleanup({ extensions: ['ts'] })],
         output: [
             {
-                file: pkg.main,
+                file: './dist/index.js',
                 format: 'cjs',
                 sourcemap: true,
             },
             {
-                file: pkg.module,
+                file: './dist/index.mjs',
                 format: 'es',
                 sourcemap: true,
             },
         ],
     }),
     bundle({
+        input: './src/lib/evaluators/index.ts',
+        plugins: [commonjs(), resolve(), esbuild(), cleanup({ extensions: ['ts'] })],
+        output: [
+            {
+                file: './dist/evaluators/index.js',
+                format: 'cjs',
+                sourcemap: true,
+            },
+            {
+                file: './dist/evaluators/index.mjs',
+                format: 'es',
+                sourcemap: true,
+            },
+        ],
+    }),
+    bundle({
+        input: './src/index.ts',
         output: {
             file: pkg.types,
+            format: 'es',
+        },
+        plugins: [resolve(), commonjs(), cleanup({ extensions: ['.ts'] }), dts()],
+    }),
+    bundle({
+        input: './src/lib/evaluators/index.ts',
+        output: {
+            file: './dist/types/evaluators.d.ts',
             format: 'es',
         },
         plugins: [resolve(), commonjs(), cleanup({ extensions: ['.ts'] }), dts()],
