+++ src/lib/Evaluators/index.ts
@@ -1,6 +1,5 @@
 /** @format */
 
-export { evaluateFindOptions } from './evaluateFindOptions';
 export { Exactly } from './Exactly.evaluator';
 export { Is } from './Is.evaluator';
 export { Like } from './Like.evaluator';
