+++ tsconfig.json
@@ -1,4 +1,6 @@
 {
+  "include": ["./src/**/*.ts"],
+  "exclude": ["!./src/**/*.ts"],
     "compilerOptions": {
       /* Visit https://aka.ms/tsconfig to read more about this file */
